initial diagrammer refactor
