+++ src/d3-utils/tables.ts
@@ -8,7 +8,7 @@ import {
 import * as d3 from 'd3';
 import { getLabel } from './styles';
 import { SelectionInfoPacket } from "../components/interfaces"
-import {DiagramField} from "../utils/diagrammer"
+import {DiagramField} from "../utils/LookmlDiagrammer/"
 import {
   isTableRowDimension,
   isTableRowMeasure,
@@ -79,27 +79,13 @@ export function createLookmlViewElement(
   });
 
   let tableTopCap = () => tableRow.append("path")
-  .attr("d", (dd: DiagramField, i: number) => i === 0 ? `M0,${TABLE_ROW_HEIGHT}
-    v-${TABLE_ROW_HEIGHT - CAP_RADIUS}
-    q0,-${CAP_RADIUS} ${CAP_RADIUS},-${CAP_RADIUS}
-    h${TABLE_WIDTH - (CAP_RADIUS * 2)}
-    q${CAP_RADIUS},0 ${CAP_RADIUS},${CAP_RADIUS}
-    v${TABLE_ROW_HEIGHT - CAP_RADIUS}
-    z
-  ` : `M0,0`)
+  .attr("d", (dd: DiagramField, i: number) => i === 0 ? `M0,${TABLE_ROW_HEIGHT} v-${TABLE_ROW_HEIGHT - CAP_RADIUS} q0,-${CAP_RADIUS} ${CAP_RADIUS},-${CAP_RADIUS} h${TABLE_WIDTH - (CAP_RADIUS * 2)} q${CAP_RADIUS},0 ${CAP_RADIUS},${CAP_RADIUS} v${TABLE_ROW_HEIGHT - CAP_RADIUS} z` : `M0,0`)
   .classed("table-row", true)
   .classed("table-row-view", (d: DiagramField) => isTableRowView(d))
   .classed("table-row-base-view", (d: DiagramField) => isTableRowBaseView(d))
 
   let tableBottomCap = () => tableRow.append("path")
-  .attr("d", (dd: DiagramField, i: number) => i === (tableData.length - 1) ? `M0,0
-    h${TABLE_WIDTH}
-    v${TABLE_ROW_HEIGHT - CAP_RADIUS}
-    q0,${CAP_RADIUS} -${CAP_RADIUS},${CAP_RADIUS}
-    h-${TABLE_WIDTH - (CAP_RADIUS * 2)}
-    q-${CAP_RADIUS},0 -${CAP_RADIUS},-${CAP_RADIUS}
-    z
-  ` : `M0,0`)
+  .attr("d", (dd: DiagramField, i: number) => i === (tableData.length - 1) ? `M0,0 h${TABLE_WIDTH} v${TABLE_ROW_HEIGHT - CAP_RADIUS} q0,${CAP_RADIUS} -${CAP_RADIUS},${CAP_RADIUS} h-${TABLE_WIDTH - (CAP_RADIUS * 2)} q-${CAP_RADIUS},0 -${CAP_RADIUS},-${CAP_RADIUS} z` : `M0,0`)
   .classed("table-row", true)
   .classed("table-row-dimension", (d: DiagramField) => isTableRowDimension(d))
   .classed("table-row-measure", (d: DiagramField) => isTableRowMeasure(d))
