+++ src/test_data/order_items_dimensions.ts
@@ -1,3 +1,3 @@
-import { DiagramMetadata } from "../utils/diagrammer";
+import { DiagramMetadata } from "../utils/LookmlDiagrammer";
 
 export const view_dimensions: any = {"exploreName":"order_items","modelName":"2020-election-betting","diagramDict":{"joinData":[],"yOrderLookup":{"0":["order_items"]},"tableData":{"order_items":[{"category":"view","view":"order_items","name":"order_items","base":true,"diagramX":0,"diagramY":190,"fieldTypeIndex":0,"diagramDegree":0,"verticalIndex":1},{"align":"right","can_filter":true,"category":"dimension","default_filter_value":null,"description":null,"enumerations":null,"field_group_label":null,"fill_style":null,"fiscal_month_offset":0,"has_allowed_values":false,"hidden":false,"is_filter":false,"is_numeric":true,"label":"Order Items Cost","label_from_parameter":null,"label_short":"Cost","map_layer":null,"name":"order_items.cost","strict_value_format":false,"requires_refresh_on_sort":false,"sortable":true,"suggestions":null,"tags":[],"type":"number","user_attribute_filter_types":["number","advanced_filter_number"],"value_format":"$#,##0.00","view":"order_items","view_label":"Order Items","dynamic":false,"week_start_day":"monday","dimension_group":null,"error":null,"field_group_variant":"Cost","measure":false,"parameter":false,"primary_key":false,"project_name":"2020-election-betting","scope":"order_items","suggest_dimension":"order_items.cost","suggest_explore":"order_items","suggestable":false,"is_fiscal":false,"is_timeframe":false,"can_time_filter":false,"time_interval":null,"lookml_link":"/projects/2020-election-betting/files/views%2Forder_items.view.lkml?line=11","permanent":null,"source_file":"views/order_items.view.lkml","source_file_path":"2020-election-betting/views/order_items.view.lkml","sql":"${TABLE}.cost ","sql_case":null,"filters":null,"diagramX":0,"diagramY":190,"fieldTypeIndex":0,"diagramDegree":0,"verticalIndex":1},{"align":"left","can_filter":true,"category":"dimension","default_filter_value":null,"description":null,"enumerations":null,"field_group_label":"Created Date","fill_style":"range","fiscal_month_offset":0,"has_allowed_values":false,"hidden":false,"is_filter":false,"is_numeric":false,"label":"Order Items Created Date","label_from_parameter":null,"label_short":"Created Date","map_layer":null,"name":"order_items.created","strict_value_format":false,"requires_refresh_on_sort":false,"sortable":true,"suggestions":null,"tags":[],"type":"date_date","user_attribute_filter_types":["datetime","advanced_filter_datetime"],"value_format":null,"view":"order_items","view_label":"Order Items","dynamic":false,"week_start_day":"monday","dimension_group":"order_items.created","error":null,"field_group_variant":"Date","measure":false,"parameter":false,"primary_key":false,"project_name":"2020-election-betting","scope":"order_items","suggest_dimension":"order_items.created_date","suggest_explore":"order_items","suggestable":false,"is_fiscal":false,"is_timeframe":true,"can_time_filter":false,"time_interval":{"name":"day","count":1},"lookml_link":"/projects/2020-election-betting/files/views%2Forder_items.view.lkml?line=17","permanent":null,"source_file":"views/order_items.view.lkml","source_file_path":"2020-election-betting/views/order_items.view.lkml","sql":"${TABLE}.created_at ","sql_case":null,"filters":null,"diagramX":0,"diagramY":190,"fieldTypeIndex":1,"diagramDegree":0,"verticalIndex":1},{"align":"left","can_filter":true,"category":"dimension","default_filter_value":null,"description":null,"enumerations":[{"label":"Below 0","value":"Below 0"},{"label":"0 to 4","value":"0 to 4"},{"label":"5 to 9","value":"5 to 9"},{"label":"10 to 19","value":"10 to 19"},{"label":"20 to 39","value":"20 to 39"},{"label":"40 to 79","value":"40 to 79"},{"label":"80 to 159","value":"80 to 159"},{"label":"160 to 359","value":"160 to 359"},{"label":"360 or Above","value":"360 or Above"},{"label":"Undefined","value":"Undefined"}],"field_group_label":null,"fill_style":"enumeration","fiscal_month_offset":0,"has_allowed_values":false,"hidden":false,"is_filter":false,"is_numeric":false,"label":"Order Items Days Since Arrival Tier","label_from_parameter":null,"label_short":"Days Since Arrival Tier","map_layer":null,"name":"order_items.days_since_arrival_tier","strict_value_format":false,"requires_refresh_on_sort":false,"sortable":true,"suggestions":["Below 0","0 to 4","5 to 9","10 to 19","20 to 39","40 to 79","80 to 159","160 to 359","360 or Above","Undefined"],"tags":[],"type":"tier","user_attribute_filter_types":["string","advanced_filter_string"],"value_format":null,"view":"order_items","view_label":"Order Items","dynamic":false,"week_start_day":"monday","dimension_group":null,"error":"Unknown or inaccessible field \"order_items.days_since_arrival\" referenced in \"order_items.days_since_arrival_tier\". Check for typos or append a timeframe to the name if the field is type \"time\".","field_group_variant":"Days Since Arrival Tier","measure":false,"parameter":false,"primary_key":false,"project_name":"2020-election-betting","scope":"order_items","suggest_dimension":"order_items.days_since_arrival_tier","suggest_explore":"order_items","suggestable":true,"is_fiscal":false,"is_timeframe":false,"can_time_filter":false,"time_interval":null,"lookml_link":"/projects/2020-election-betting/files/views%2Forder_items.view.lkml?line=46","permanent":null,"source_file":"views/order_items.view.lkml","source_file_path":"2020-election-betting/views/order_items.view.lkml","sql":"${days_since_arrival} ","sql_case":null,"filters":null,"diagramX":0,"diagramY":190,"fieldTypeIndex":2,"diagramDegree":0,"verticalIndex":1},{"align":"right","can_filter":true,"category":"dimension","default_filter_value":null,"description":null,"enumerations":null,"field_group_label":null,"fill_style":null,"fiscal_month_offset":0,"has_allowed_values":false,"hidden":false,"is_filter":false,"is_numeric":true,"label":"Order Items ID","label_from_parameter":null,"label_short":"ID","map_layer":null,"name":"order_items.id","strict_value_format":false,"requires_refresh_on_sort":false,"sortable":true,"suggestions":null,"tags":[],"type":"number","user_attribute_filter_types":["number","advanced_filter_number"],"value_format":null,"view":"order_items","view_label":"Order Items","dynamic":false,"week_start_day":"monday","dimension_group":null,"error":null,"field_group_variant":"ID","measure":false,"parameter":false,"primary_key":true,"project_name":"2020-election-betting","scope":"order_items","suggest_dimension":"order_items.id","suggest_explore":"order_items","suggestable":false,"is_fiscal":false,"is_timeframe":false,"can_time_filter":false,"time_interval":null,"lookml_link":"/projects/2020-election-betting/files/views%2Forder_items.view.lkml?line=5","permanent":null,"source_file":"views/order_items.view.lkml","source_file_path":"2020-election-betting/views/order_items.view.lkml","sql":"${TABLE}.id ","sql_case":null,"filters":null,"diagramX":0,"diagramY":190,"fieldTypeIndex":3,"diagramDegree":0,"verticalIndex":1},{"align":"left","can_filter":true,"category":"dimension","default_filter_value":null,"description":null,"enumerations":[{"label":"Yes","value":"Yes"},{"label":"No","value":"No"}],"field_group_label":null,"fill_style":"enumeration","fiscal_month_offset":0,"has_allowed_values":false,"hidden":false,"is_filter":false,"is_numeric":false,"label":"Order Items Is Sold (Yes / No)","label_from_parameter":null,"label_short":"Is Sold (Yes / No)","map_layer":null,"name":"order_items.is_sold","strict_value_format":false,"requires_refresh_on_sort":false,"sortable":true,"suggestions":null,"tags":[],"type":"yesno","user_attribute_filter_types":["string","advanced_filter_string"],"value_format":null,"view":"order_items","view_label":"Order Items","dynamic":false,"week_start_day":"monday","dimension_group":null,"error":"Unknown or inaccessible field \"order_items.sold_raw\" referenced in \"order_items.is_sold\". Check for typos or append a timeframe to the name if the field is type \"time\".","field_group_variant":"Is Sold (Yes / No)","measure":false,"parameter":false,"primary_key":false,"project_name":"2020-election-betting","scope":"order_items","suggest_dimension":"order_items.is_sold","suggest_explore":"order_items","suggestable":true,"is_fiscal":false,"is_timeframe":false,"can_time_filter":false,"time_interval":null,"lookml_link":"/projects/2020-election-betting/files/views%2Forder_items.view.lkml?line=35","permanent":null,"source_file":"views/order_items.view.lkml","source_file_path":"2020-election-betting/views/order_items.view.lkml","sql":"${sold_raw} is not null ","sql_case":null,"filters":null,"diagramX":0,"diagramY":190,"fieldTypeIndex":4,"diagramDegree":0,"verticalIndex":1},{"align":"right","can_filter":true,"category":"measure","default_filter_value":null,"description":null,"enumerations":null,"field_group_label":null,"fill_style":null,"fiscal_month_offset":0,"has_allowed_values":false,"hidden":false,"is_filter":false,"is_numeric":true,"label":"Order Items Average Cost","label_from_parameter":null,"label_short":"Average Cost","map_layer":null,"name":"order_items.average_cost","strict_value_format":false,"requires_refresh_on_sort":false,"sortable":true,"suggestions":null,"tags":[],"type":"average","user_attribute_filter_types":["number","advanced_filter_number"],"value_format":"$#,##0.00","view":"order_items","view_label":"Order Items","dynamic":false,"week_start_day":"monday","dimension_group":null,"error":null,"field_group_variant":"Average Cost","measure":true,"parameter":false,"primary_key":false,"project_name":"2020-election-betting","scope":"order_items","suggest_dimension":"order_items.average_cost","suggest_explore":"order_items","suggestable":false,"is_fiscal":false,"is_timeframe":false,"can_time_filter":false,"time_interval":null,"lookml_link":"/projects/2020-election-betting/files/views%2Forder_items.view.lkml?line=82","permanent":null,"source_file":"views/order_items.view.lkml","source_file_path":"2020-election-betting/views/order_items.view.lkml","sql":"${cost} ","sql_case":null,"filters":null,"diagramX":0,"diagramY":190,"fieldTypeIndex":0,"diagramDegree":0,"verticalIndex":1},{"align":"right","can_filter":true,"category":"measure","default_filter_value":null,"description":null,"enumerations":null,"field_group_label":null,"fill_style":null,"fiscal_month_offset":0,"has_allowed_values":false,"hidden":false,"is_filter":false,"is_numeric":true,"label":"Order Items Count","label_from_parameter":null,"label_short":"Count","map_layer":null,"name":"order_items.count","strict_value_format":false,"requires_refresh_on_sort":false,"sortable":true,"suggestions":null,"tags":[],"type":"count","user_attribute_filter_types":["number","advanced_filter_number"],"value_format":null,"view":"order_items","view_label":"Order Items","dynamic":false,"week_start_day":"monday","dimension_group":null,"error":null,"field_group_variant":"Count","measure":true,"parameter":false,"primary_key":false,"project_name":"2020-election-betting","scope":"order_items","suggest_dimension":"order_items.count","suggest_explore":"order_items","suggestable":false,"is_fiscal":false,"is_timeframe":false,"can_time_filter":false,"time_interval":null,"lookml_link":"/projects/2020-election-betting/files/views%2Forder_items.view.lkml?line=88","permanent":null,"source_file":"views/order_items.view.lkml","source_file_path":"2020-election-betting/views/order_items.view.lkml","sql":"order_items.count","sql_case":null,"filters":null,"diagramX":0,"diagramY":190,"fieldTypeIndex":1,"diagramDegree":0,"verticalIndex":1}]}}}
